Began Work on marker controller
md5sum a.out
78e4c7b82063c0250ddc2363d22d3a2c  a.out

The marker controller has a basic function that is called from the
network layer and simply returns an empty JSON object and a 503 error.
The first step is to retrieve the pins from the marker database and then
send them out to the client.

After that we'll refactor the pins to use API spec version 3 instead of
2 and make the GET work with that as well.