Skeleton of Marker Controller Function
md5sum a.out
de399708c6e61843aeaeca2fcc4fd445  a.out

The beginnings of the 'routing' function for the marker's API endpoint
is in progress. It can distinguish between the various methods but
doesn't actually call any methods yet.

It also does not use heap storage for it's buffer string (due to the
nature of the marker's themselves) and malloc's instead. Eventually I'd
like to have all strings malloced and realloced for when I need more
memory that way I don't get buffer overflows and stack smashing.