Delete Marker Endpoint functional
md5sum a.out
5ed08291c238644f7a81454c075f80b6  a.out

Added marker delete database function to handle deleting. Sensible
defaults for the values for latitude and longitude are NULL. This way
we can check for this and alter the query we send to the database based
on the parameters actually present. The only downside is the error
handling code for possible lack of memory grows pretty fast when you're
malloc-ing everything.

Defaults for the offsets if not present are an arbitrary chosen 0.5 and
is defined in controllers/markers.h

Added parameter checking for the offsets so that non-numeric values are
not allowed.