Began adding addressed field to markers
md5sum a.out
7e62685632b6dcc18391e0d8bd11275d  a.out

Version 3 of the API spec dictates that the markers have an 'addressed'
field that is used to inform the client if whatever information the
marker represents has been addressed in some way by other users.

The addressed field has been added to the SQL setup and to the model
object itself. The next step is to add it to the database queries and
the json strings. Then verify that memory sizes for the strings are
still within limits and we don't smash the stack.