Marker Controller Responds to Addressed
md5sum a.out
7eff15b6d15eafa2eec81adaff07cedb  a.out

The addressed method of the marker endpoint is functional. It is not
covered by any unit tests or database tests, I have tested it myself but
that's not replicatable by other people since they're not me.

Also, if you're wondering why sometimes the server might say address in
use after closing it's because of this:
http://stackoverflow.com/a/5106755/1808164

It may be good to use SO_REUSEADDR for the socket so we don't get this
problem.