No Parameter GET against markers functional
md5sum a.out
19f3e24a5105a1d378623cfa667ce7ed  a.out

Created database queries to perform joins between markers and comments
to get the appropriate data to create the hybrid-view that the marker
endpoint returns.

Created JSON function to create the 'hybrid' view mentioned from a
comment and a pin. This does mean that I retrieve and construct TWO
lists, one of comments, one of markers and that could be considered
wasteful since I don't use all of either of them. But until it poses a
problem, I won't optimize and create a different struct, you never know
when one might need that extra information.

Added functionality to the controller to respond to a default no
parameter request against the marker controller.