Added longitude filtering to marker endpoint
md5sum a.out
6e41b2d33499fc7e019da4726073fecb  a.out

Added support for just longitudinal filtering. The code for the database
function is literally the exact same as the latitude besides the actual
query used against the database.

I think that refactoring those two functions to be one function and then
two smaller ones that just decide the query before calling the common
functionality might be a good idea.