Added Full Filter. Found bug with Decimals
md5sum a.out
3fa5b96dc991f0501df5c6e0325255c9  a.out

Added full filter search for longitude and latitude at the same time
against the marker endpoint. It is _mostly_ working. There remains one
more little fix that needs to get done (within Decimal)

Refactored the by longitude and by latutitude functions into a more
general by coordinate one that takes a query as a parameter.

The issue with the Decimal occurs if you try to do the following:

negative number - positive number.
Like "-44 - 1", this will get you -44 which is not correct.