Decimal Refactor.
md5sum a.out
e4b054ac72cc4bd2e020c1cc977a8583  a.out

Major Patch. Refactored Decimal struct into a simple define, so we can
change the precision by simply using either long double or double
defined as Decimal, then the conversion function is defined as
strToDecimal and the correct formatter is defined as DecimalFormat. All
of this can be found in decimal.h, and if one wanted to switch to an
'exact' library, they'd merely substitute their own decimal functions in
place of the current ones.