Heatmap PUT Endpoint functional
md5sum a.out
54292c3ca689caac5b6bdf09f3b391f3  a.out

Sending a PUT request to the endpoint now functions correctly and
creates a heatmap record in the database.

Supports Error messages on faulty requests to the endpoint, although the
JSON parsing I'm doing enables the JSON to have slightly poor structure
and the : between pairs may be left out and it will still parse out the
values.