Heatmap Default GET working. Raw/Precision Support
md5sum a.out
6ba2232a645aad32f2b7b833bff87688  a.out

Precision support is working, truncates the precision correctly. The Raw
parameter is working as well, current it's set to true, but could easily
change to default to false. Returns the percentage back as an integer
between 0-100

The request to get all the heatmap points regardless of offsets or
degrees specified is working.