Removed database queries from logging.
md5sum a.out
73fb5f2fb75b288291aa5a2d3e8f3df6  a.out

I may add the database logging back in, or perhaps have it log to a
different thread. It would be interesting to set up macros to define
whether logging is turned on or not, so that when you compile the
program you can deteremine the amount of time it spends logging,
something like:

 #define LOG_QUERIES 1

then at startup:

 #ifdef LOG_QUERIES
    .. open file descriptor and whatnot and save as read only global.

of course that leads to concurrency issues though, so we'd need a
concurrent queue of some kind.