POST Method of Reports Endpoint.
md5sum a.out
f3c117be9b47902beb826eb03cfbcd30  a.out

Implemented POST Method, only thing missing is the stack trace because
during the creation of the model I forgot to include the stack trace
field. This will be implemented soon.

Posting according to the specification creates an entity in the database
with the appropriate fields filled out (besides stackTrace of course)