Removed crash code
md5sum a.out
b58c0bedc6716031f281e35840bde341  a.out

Removed thread joining code that was a mistake to have in the first
place. The code was unneccesary because the join is done inside the loop
anyway and for some reason I was blind to the join not being within the
inner for loop despite how little sense it would have made if it were
otherwise.