Fixed NETWORK_LOGGING Levels. Found shutdown bug
md5sum a.out
2f426911da61f02dc11eaeed1e26bfc3  a.out

Fixed conditional for determining network logging level if not set,
using an OR was a slipup that is now fixed. Network logging is now
enabled, which also enabled me to find the source of the warning about
the bad file descriptor.

A final request is parsed after the stop command is sent. It attempts to
send the heartbeat out to an invalid file descriptor. This was never a
problem during testing because all the threads ended cleanly and we
never sat around waiting for the network to get a client. We can solve
this with polling or a simple check against the fd.