Removed Memory Tables.
md5sum a.out
851d18e8a18b3bd6bbba43f6397ad027  a.out

Memory Tables handle updates through Full Table Locks, which means that
in a highly concurrent environment with lots of updates they'd be
unstainable.

Comments and Markers, while not the focus of the application, will
probably be updated enough that using a M. Table wouldn't help that
much -- and the benefits of a very fast read, might be negated by the
problem of updating the information.

If we were to allow for an update period, we could use a scheduled job
that would rebuild the memory tables every X seconds/minutes. This would
mean we'd be sending out stale data during those X seconds/minutes but
we'd be sending it out very fast. The rebuilding process might not even
be that bad considering it would be a truncation of the memory table and
then a single repopulation.

For now, the cache tables will be removed.