Fixed Testing framework linking
md5sum a.out
c33fc48e6738ce3efb99f8b620cf33bb  a.out

Due to strmap's inclusion into the JSON helper, the linking code the
test framework had to be fixed. Speaking of the testing framework, I
would really like to extend it to actually test the controllers well and
to make sure their methods err when they're supposed to do and what not.
The first step towards this is refactoring the comments controller to
behave like the others.

Besides no_mem errors, the controller sub procedures handling each of
the HTTP verbs should be possible to test by passing them mock buffers.
The delegation function for each controller should be able to be mocked
by passing a fake thread data struct with a msg buffer as well as a fd
of STDOUT_FILENO in order to have the buffer be sent out to the
terminal.