Refactored Comments POST, closed STD I/O
md5sum a.out
b9e8fcb4005ce8b3de1f91f0057e2296  a.out

Refactored POST method to return proper error messages from within
sub routine in a similar fashion to the other controllers. Also fixed
not being able to check if a comment message was beyond the character
limit. The problem was that I had limited the size of my buffers to be
the limit itself, and so checking if something inside the buffer was
beyond it was impossible to do via strlen.

Closed the standard file descriptors from the main function at the end
so that Valgrind doesn't report the 3 open, even though there is nothing
wrong with leaving the 3 STD File descriptors open since the main
program cleans up, it is nice to see nothing coming from the output of
valgrind.