mList implementation
md5sum a.out
72ab212d77885b7aeeea7ac7781f80da  a.out

mList is a simple forward linked list designed to store any data (via
casting to void*). It's intent is to be used within the heatmap
controllers PUT action to construct a list of model objects that will be
inserted, this way we can parse the data within the request without
opening and closing the database connection each time.

Thanks for @primehunter326 for pointing out that detail I glossed over
in my implementation.

Added test to mlist, working correctly.