Wrapped Heatmap PUT in Transaction
md5sum a.out
4d9be90b77da1dcf56f4fd4a1966b095  a.out

Wrapped heatmap put operation in transaction in order to abort all added
points if any part of the request is invalid. The big question is,
should we err if no points are given at all, or bring up an error if the
transaction aborts?

The API specification isn't clear on how to handle an empty list.