Fixed test compilation, Macro for non-null freeing
md5sum a.out
151d9bf7d03672f5c0f6e39d52dde290  a.out

Created a simple macro to free the pointers associated with latitude and
longitude degrees and offsets. Should probably undef it at the end of
the file so it is not used anywhere else, not that it could really be
used anywhere else since it's just short hand for playing a little code
golf.