I/O Polling with Select before accept
md5sum a.out
58006242656c178a65879af956d8e29f  a.out

Instead of blocking with accept, we now check the socket descriptor for
any data with a call to select before opening a connection. This means
that we might be able to have the socketfd become local again and not
closed off by the  Signal interupt, which seems like a hack to me. It
would be nice to have a more traceable execution flow.