Server Socket File Descriptor localized
md5sum a.out
c814cef5581ba87f7364ebf8a0c05872  a.out

Because of the introduction of select in the last commit, we no longer
need to have a global scope for the server sockets file descriptor. The
control flow of the program is now more readily readable and
understandable.