Refactored Driver, routing specification test
md5sum a.out
7c0042d6b827e86fe52efb4e5beab241  a.out

Modified green-serv.c to not hold the main file so that we can build up
the entire environment without conflicting with a second main function.

Added routing.c which does pretty much the same thing as the routing
test except that we will modify it to check that all endpoints of the
API will delegate to their correct controllers.

Note that we do not check that the response is correct, merely that the
routing is working properly.