Updated GET comments text, added basic getheatmap
md5sum a.out
d503e4ca9b18de9089d36fcb6a2988a8  a.out

Updated the getcomments test text to use the word request instead of
plain, plain being the leftover variable name I had before.

Created spec test of GET heatmap endpoint. 5  Failing tests! This is
exactly what this spec testing is designed to find. This is a good
thing, will continue creating the basic status code tests in order to
catch these for each of the endpoints.

Within the spider testing of the Python API over in the GreenUp
repository, the spider verifies that the responses have a set format,
which is all well and good, but is a massive pain to update. Checking
the C API's responses for consistency in the keys might take a small
amount of effort, but I'm thinking that a simple JSON parser that
creates named tokens, and then checking those tokens against a list of
expected ones should be robust.