Delete Debug SPEC Test.
md5sum a.out
0dc2ba2e95b9019dbe957f29178dbf2c  a.out

Reminder, the origin portion of the url is a non-hashed version. The
server stores a hash version for privacy reasons, but the request itself
just needs to send along the regular one.

Last 3 tests for delete spec failing for debug endpoint. This is going
to be fun once I finish writing all my tests and get around to coding
until they all pass with flying colors.