PUT Pins Spec check test
md5sum a.out
0dc2ba2e95b9019dbe957f29178dbf2c  a.out

Specification test for the PUT endpoint of the pins/Markers endpoint.
Like a few of the others, the conversions of id's and ensuring
non-numeric values are not allowed is something that is a failing test
right now but will be addressed later on after all the tests are
written.