PUT Heatmap Fix.
md5sum a.out
e9bbcd4d6773af0f590a3ff91fcb8691  a.out

Couldn't help but fix the heatmaps a bit for the put test, now just
getting a database error so its fine and I'll leave it at that (foreign
key constraint)

Updated Macro to spit out the result of the stringToReturn so that
testing is easier to see what might have gone wrong. Becuase of seeing
all the same errors returned I knew there was something up with the JSON
parsing. Fixed, and will need to propagate this change out when doing
test fixes later on.