All Specification Tests Passing.
md5sum a.out
76e1fb055fe11640f0832fb312586003  a.out

Specification tests that are ensuring that the status codes returned are
correct are now passing. The API has much more validation than before to
ensure numericality of parameters (although some NaN checks would be
good in a few places)

Load Testing and other tests should be tested soon. As well as response
checking. Response checking will probably be delegated to a language
more suitable for JSON parsing than C. As JSON is object oriented and C
is far from that.