Add Report Types to Model/Database. (Breaking)
md5sum a.out
0041a415b9cad75ed2e8bff15bbe6008  a.out

Added types to the reports object and database interaction. No
controller actions are being taken by it yet, and the creation for the
reports has not been updated yet either. However the field is there for
early adoption.

Note that it is a breaking change to the reports endpoint if you do not
migrate your database correctly. The statement to issue is simply:

ALTER TABLE report ADD COLUMN r_type CHAR(20) NOT NULL DEFAULT "INFO"

Intended plan for report types is to use a case insensitve strpos
against the types constant from the report header.