Refactored Markers to use dynamic memory response
Switched net.c's stringToReturn to be a char** in order to pass the
string into the function and be able to modify and resize it. Updated
markers to handle and test this.

Now able to bump the marker limit up again to 100 becuase we can handle
creating the JSON without truncating the result.

Updated marker spec tests to handle this correctly.

Note that the define for the err label and return had to be swapped in
the marker controller in order to handle and test this. However we shall
soon update the other controllers to handle this as well and then that
 #undef-ing will not be neccesary.