Updated stringToReturn type to all controllers.
md5sum a.out
5ca95f3e7a6d9660afa219159da1ff78  a.out

Modified char * -> char ** for all controllers so that if we need to
update them to swap memory in order to support a non-truncated response,
we can do so. Things like GET on resources that return lists of JSON objects
need to be updated to handle this.