Refactor GET Comments to use dynamic memory
md5sum a.out
0aaa06166c1ccd706e41f8fdff1ac3aa  a.out

Since the GET endpoint of comments is a list of indeterminant size we
need to use dynamic memory instead of the stack space we were alloting
before. This has been done and tested to work by lowering the request
size to a small number and ensuring that the memory was swapped up for
higher space.