Fix for secondary issue in #169
md5sum a.out
7c6f240635481feaf6e437444d6dd16c  a.out

Updated ALL comments query to use LEFT join, I had it up to date for the
other types of comments queries but missed the most important one it
looks like.

Also updated Makefile to include mysql_config --cflags in controller
calls so that it doesn't rely on system dependent paths to resolve the
include. (This is a result of my having to update mysql from 5.5 to 5.6
on my local machine and foobaring my system with it, see the
stackoverflow here:
stackoverflow.com/questions/22069856/cant-find-mysql-mysql-h-after-upgrading-to-5-6 )