Update to header to be cross-machine
md5sum a.out
04af7425cc993750a9cd1384bab48054  a.out

Added a header flag to switch the mysql include to be one for a messed
up system like mine or for one that's not so broken.

Updated Makefile to use a shell execution of mysql_config for the --libs
but I can't do that yet for the --cflags since my include directory
isn't configured right on my local machine.