Started working on create script.
Create script is going to be a simple C script to create a config.h
file based off of a key=value file.

It will print to stdout the contents of the .h file, and then we'll
write a make file command that will call the script to generate the
headers and switch them out and compile multiple versions of the api for
deploment.