Create Configure script.
Created small script to spit out a config.h file based on a .gs.conf
file or a filename passed to it. Plans are to use it within a make
command to create config.h's build the application with different
configs and then compile them out to somewhat logically named
executables.